requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# CLIP and Embeddings
git+https://github.com/openai/CLIP.git
open_clip_torch>=2.20.0

# Vector Search
faiss-cpu>=1.7.4
chromadb>=0.4.0

# VLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web & API
fastapi>=0.100.0
streamlit>=1.25.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
pyyaml>=6.0
kaggle>=1.5.16

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
